Rev3/requirements.txt
langchain
langchain-openai
langchain-community
openai
pandas
numpy
chromadb
sentence-transformers
python-dotenv
colorama
fastapi
uvicorn[standard]
python-multipart
jinja2